debug option to use different dtype

+++ aphin/layers/system_layer.py
@@ -41,7 +41,7 @@ class SystemLayer(tf.keras.layers.Layer):
         **kwargs : dict
             Additional arguments for the tf.keras.layers.Layer base class.
         """
-        super(SystemLayer, self).__init__(**kwargs)
+        super(SystemLayer, self).__init__(dtype=dtype, **kwargs)
         self.dtype_ = dtype
         self.r = r
         self.n_u = 0 if n_u is None else n_u
